Move MultiMap to Metamorph and rename it to Maps
The `MultiMap` interface is only used by Metamorph. It is made part of
the morph API to provide access to the maps defined in a Metamorph
script. To reflect this, it is renamed to `Maps`.

The `MultiHashMap` implementation is removed as the `Maps` interface is
directly implemented by the `Metamorph` class.