Remove test case for JndiSqlMap
The test case attempted only to get a data source but did not actually
check the functionality of the map. As test case does not work any more
due to the removal of the `getDataSource()` method in the previous
release, the test completely removed. This also allows us to remove the
test dependency `simple-jndi` which is quite old and currently
unmaintained.