Metamorph: Rewrite open file logic in FileMap
`FileMap` used `ResourceUtil#getStream(String)` for opening the map
data files. This is the only use of a method from the `util` package in
Metamorph's functions, collectors and maps. By removing this dependency
it becomes easier to split Metamorph in a base module and a core
functionality module as planned in the Metafacture 5.0.0 release.

The rewrite duplicates the code in `ResourceUtil#getStream(String)`.
However, this is only a temporary issue. In a future release the `files`
attribute will be revised to accept URL patterns as defined by the
xinclude standard (see issue #261).