Metamorph: Separate internal and external exceptions
In order to make the Metamorph API independent of the Metafacture
framework `MorphDefException` and `MorphException` are replaced by two
new exceptions (`MorphBuildException` and `MorphExecutionException`).
These do not extend `MetafactureException` to achieve the independence
of the framework. As a consequence they are considered an internal part
of Metamorph and its extensions.

If one of these exceptions occurs Metamorph wraps it in a
`MetamorphException`. This exception extends `MetafactureException`. It
is this exception that is visible to users of Metamorph.

From a user's perspective this commit replaces `MorphDefException` and
`MorphException` with `MetamorphException`.