Do not wrap IOExceptions in ResourceUtil
Some of the methods in `ResourceUtil` wrap `IOException`s in
`MetafactureException`s. These methods are changed to directly throw
`IOException`s. This leads to a more constistent interface as now all
methods in `ResourceUtil` handle exceptions in the same way.

Not wrapping the exceptions has the advantage that the caller can decide
what the appropriate wrapper exception is. This may depend on the
context in which the `ResourceUtil`s are used. For instance, the failure
to load a properties file might be a user error if the user provided the
file name but it may be a system error if the properties are packaged as
a resource in the application.

By not using `MetafactureException` anymore we can remove the dependency
of `ResourceUtil` on the `framework` package.