Refactor ObjectFactory
The `ObjectFactory` combined a directory of classes which could be
instantiated dynamically using custom names with utility functions for
dynamically instantiating and initialising objects. This commit extracts
these utility functions and introduces a `ConfigurableClass` class for
handling the dynamic instantiation and initialisation of classes.

This refactoring changes the interface of the old `ObjectFactory` as a
number of functions have been moved to `ConfigurableClass` and
`ReflectionUtil`.