Refactor Metamorph tests to not use metamorph-test
Converts the test cases for Metamorph that were written in
metamorph-test xml to standard JUnit based tests. This makes them easier
to understand and offers more flexiblity when checking the test results.
Additionally, it breaks the circular dependency on the `test` package
(this is the package containing the metamorph-test implementation).