Move metamorph-test xml files to resources
The metamorph-test xml files used to be placed in the same folder as the
Java class which linked them into the JUnit process. This required a
special Maven configuration in order to ensure that the test resources
from the src directory were copied into the target directory. As there
are only two metamorph-test based test files left after converting the
Metamorph tests to standard JUnit tests, it is not reasonable to keep
this special treatment of resources in the src folder. This commit moves
the remaining metamorph-test files to the resources folder and removes
the Maven configuration for copying the resources.