Rename package root and convert to multi-module project
The package root is changed from `org.culturegraph.mf` to
`org.metafacture`.

Additionally, the project is converted into a multi-module project that
splits the former `metafacture-core` artifact into several smaller
artifacts. This allows users of Metafacture to reduce the number of
external dependencies pulled by Metafacture to the ones which are
actually required. It also helps to minimize dependencies between
Metafacture modules and makes the remaining dependencies more epxlicit.