Refactor TarReader to improve stream handling
The `try` statements in `TarReader` are replaced with try-with-resource
statements to ensure we are not leaking any resources.

Individual archive entries are no longer read completely into a string
before calling the downstream module. Instead of this, the archive
stream is passed directly to the downstream module. It is only wrapped
in a small adapter class which works around the special semantics of the
`close()` method in `ArchiveInputStream`.