Unescape string literals in Flux with Antlr
The Flux lexer grammar defines the escape sequences allowed in strings.
It makes sense to perform the unescaping directly during scanning using
Antlr actions instead of relying on Apache commons-lang to parse the
string literals again.

This approach has three advantages: First, it allows us to remove the
commons-lang dependency as this was the only remaining use of the
library in Metafacture. Second, it solves the problem that the
commons-lang library does not understand all escape sequences that are
allowed by the Flux grammar (commons-lang does not support octal escape
sequences while Flux does). And thirdly, the string escape rules are now
dealt with in one place instead of two.