Add ForwardingStreamPipe to framework helpers
The `ForwardingStreamPipe` is a convenience class for writing stream
modules which only want to handle some stream event types. The default
implementations of the methods for handling the stream events simply
forward the event to the downstream module. This behaviour is different
from the behaviour implemented by `DefaultStreamPipe` which discards all
unhandled events.