Migrate build system from Maven to Gradle
The main motivation for migrating to Gradle is the build duration. With
the introduction of the new multi-module project structure build times
of the Maven-based build increased heavily. Tests showed that Gradle can
build the project much faster.

In addition to reducing build times, Gradle allows to better minimize
code duplication in build scripts: Maven's solution for centralising
build configuration in a multi-module project are parent modules. The
disadvantage of this solution is that the parent modules need to be
published along with the actual artifacts.  These additional
publications, which are technically only required for building the
artifacts, are a source of confusion for users.

Other than Maven Gradle clearly separates the organisation of the build
scripts from the published dependency information. This enables us to
coveniently organise the build scripts while at the same time only
publish the actual artifacts.

Gradle's philosophy for configuring a build also opens new options for
organising the release process.