Fix failing build when not executing up to package
metafacture-runner attempted to unpack the metamorph artifact in the
`generate-resources` phase. However, if Maven was invoked with a
lifecycle phase before the the `package` phase, the metamorph artifact
does not exist in the reactor and the build fails. This commit moves the
unpacking of the metamorph artifact into the package phase to solve
the problem.