Fix missing name and description in POM
The subprojects configuration in the `build.gradle` of the root project
is executed before the build scripts of the subprojects are executed.
Since the name and the description for the POM is set in the build
scripts of the subprojects this information is not yet available at the
time when the POM is configured in the root build script.

This commit fixes this by postponing the POM configuration to the
after-evaluation phase of the build.