Do not allow information separators in record label
While it is not explicitely mentioned in the ISO2709 standard, it does
seem to be a sensible to use information separators in the
implementation codes and system chars area of the record label.

This commit adds the relevant checks to the record label.

(cherry picked from commit 9bcc8fc1a876eb81900d8a1d9a3b831fd0c92d1b)