Add ObjectThreader for multithreading using pipelines
The ObjectThreader divides incoming objects and passes them to receivers which,
in conjunction with an ObjectPipeDecoupler, are running in their own threads.
The receivers can be added like Tee receivers. Thus, a multithreaded object pipeline
is created.

- change methods of DefaultTee to not be final

The ObjectThreader makes use of the DefaultTee and overrides two methods, so
they musn't be final.

See hbz/lobid-resources#967.