Add ObjectThreader for multithreading using pipelines
The ObjectThreader divides incoming objects and passes them to receivers,
each internally plumbed with an ObjectPipeDecoupler.
Thus, a multithreaded object pipeline is created.
The receivers are to be added like every Tee receiver.

- change methods of DefaultTee to not be final
- add test

The ObjectThreader makes use of the DefaultTee and overrides two methods, so
they musn't be final.

See hbz/lobid-resources#967.