Make MARCXML namespace for record elements configurable.
Not all MARCXML sources specify the required namespace URI (notably,
Alma General Publishing). By allowing the expected namespace to be
configured, or even skipped (by setting it to `null`), this will
resolve #330.

The record element is the only element that is restricted by namespace;
however, the original commit [1] doesn't provide any context as to why
the change was introduced.

[1] https://sourceforge.net/p/culturegraph/code/1507/