Allow to pass data through metamorph
With the new keyword "_elseAndPassEntityEvents" (set with
<data source="_elseAndPassEntityEvents" /> ) the known "_else"
is triggered AND entity events for these "_else" sources are fired.
With this, data can be passed through metamorph. All "_else" data
are handled in receivers like all the other data handled by morph rules.

Data which is handled by metamorph rules will NOT be passed through
(hence the aptly named "_else"). If you want to use data in the morph
AND pass it through, you have to add an explicit rule for this, as usual.

See #107.