Encode top-level MARC record leader as proper XML element instead of control field.
The XML decoder (not knowing anything about MARC) emits the leader element as a top-level literal. This is in contrast to the MARC21 decoder which wraps it in an entity.

The latter seems to be a historical artifact, originating from the removal of the `splitLeader` setting in `Marc21Decoder` (6d04d69) and the subsequent introduction of the `emitLeaderAsWhole` setting (76eb9fd).

Fixes #336.