Fix "setreplace" using a FileMap (#381)
"AbstractReadOnlyMap" didn't allow access to the full contents of Maps. It allowed
only access by giving a key. While this makes sense when e.g. querying (big) databases
this broke the "setreplace" function where the whole Map is loaded first to be able
to replace (parts of) the input string.