Call LineRecorder receiver's closeStream() (#433)
Under some circumstances (e.g. if the JVM is not stopped after processing)
the streams of receivers connected to LineRecorder are not closed, resulting
in empty or missing output data.

By switching from "implements ObjectPipe" to "extends DefaultObjectPipe" in LineRecorde
the "DefaultSender" is used which implements the Sender interface. This ensures the
proper closing of streams.

- complement test