Throw RuntimeException in FluxParser (#421)
If a `RecognitionException` occurs in `FluxParser` it was catched. So callers had
no chance to be aware of the Exception.
As it's not possible to create a new `RecognitionException` with a customized
error message (and throw this new `RecognitionException`) and also to be
backwards compatibel (`RuntimeExceptions` are referred to as unchecked exceptions
and the code would compile just as before) no checked Exception is thrown but
a `RuntimeException` with the customized error message.