Throw ANTLR errors (#421)
This guarantees (more) exceptions to be thrown. Note that it isn't sufficient to
just add a default (or: minimal) catch handler to all expressions, e.g. adding
that to the 'atom' expression would result in an not detailed enough message.

The overriding of 'recoverFromMismatchedToken' guarantees the throwing of an
exception when attempting a recover (before that would be reported but not thrown (see
https://github.com/metafacture/metafacture-playground/issues/84 )).

- use 'FluxParseException' instead of 'RuntimeException'
- add some more tests