Compute valid top-level MARC record leader. (#524)
MarcXmlEncoder acts as a wrapper. It makes use of Marc21Encoder and Marc21Decoder to ensure proper MARCXML, especially regarding the leader. Also - in contrast to MarcXmlEncoder - the record ID (field 001) is mandatory.

The outer class is responsible for routing the stream into the appropriate pipe, the inner class contains the actual (previous) implementation.

Co-authored-by: Pascal Christoph <pascal.christoph@hbz-nrw.de>