Add how to set an empty namespace (#552)
The MARC21 namespace is checked by default.
Sometimes there is MARC21 data without the proper namespace or none namespace
at all. To make the handler run in such cases the namespace must be explicitly
set as parameter to match the namespace used in the data.

The @Description describes how to set an empty namespace if the namespace is
missing. Ignoring the namepace at all can be done only via JAVA-API by setting
it to null.